Extract sanitizers in `service_providers` directory
